# [backend](cuda): faster uncontiguous concat