- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High

Faster implementation of uncontiguous concat on cuda.
1. the Debug and Release build ci test is passed.
2. test-backend-op is passed, I also add one performance test, but since it may not be a real example,  I just post the result in the experiment.
3. the first uncontiguous concat kernel in deepseek-v2 is faster than master. However, the total performence is lower than master, which makes me a little confused. Due to variations in the running results of llama bench, anybody can run test and discuss the result.

**Performance experiment：**
**back-end-ops:**

test case
`test_cases.emplace_back(new test_concat(GGML_TYPE_F32, {512, 1024, 5, 1}, 1024, 0, 1));`

my implementation:
```
Backend 1/2: CUDA0
  Device description: NVIDIA GeForce RTX 3060 Laptop GPU
  Device memory: 6143 MB (5122 MB free)

  CONCAT(type=f32,ne_a=[512,1024,5,1],ne_b_d=1024,dim=0,v=1):                   5004 runs -   211.03 us/run -   120830 kB/run -  546.04 GB/s
  Backend CUDA0: OK
```

master:
```
  Device 0: NVIDIA GeForce RTX 3060 Laptop GPU, compute capability 8.6, VMM: yes
Testing 2 devices

Backend 1/2: CUDA0
  Device description: NVIDIA GeForce RTX 3060 Laptop GPU
  Device memory: 6143 MB (5122 MB free)

  CONCAT(type=f32,ne_a=[512,1024,5,1],ne_b_d=1024,dim=0,v=1):                   3336 runs -   322.92 us/run -   120830 kB/run -  356.84 GB/s
  Backend CUDA0: OK
``` 

**llama-bench:**

command：
`~/program/forked/llama.cpp/build/bin$ ./llama-bench -m ~/program/forked/DeepSeek-V2-Lite-Chat.IQ1_S.gguf`

**the first uncontiguous concat kernel：**
my implementation:
![image](https://github.com/user-attachments/assets/82b1a9bd-0695-442c-9f58-3d133b2e246b)

master:
![image](https://github.com/user-attachments/assets/1adf0e50-0d26-4d4e-859f-d6dcba3625d5)

**the overall result:**
my implementation:
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| deepseek2 16B IQ1_S - 1.5625 bpw |   4.65 GiB |    15.71 B | CUDA       |  99 |         pp512 |        768.63 ± 8.81 |
| deepseek2 16B IQ1_S - 1.5625 bpw |   4.65 GiB |    15.71 B | CUDA       |  99 |         tg128 |         21.84 ± 3.75 |


master:
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| deepseek2 16B IQ1_S - 1.5625 bpw |   4.65 GiB |    15.71 B | CUDA       |  99 |         pp512 |        804.11 ± 4.73 |
| deepseek2 16B IQ1_S - 1.5625 bpw |   4.65 GiB |    15.71 B | CUDA       |  99 |         tg128 |         20.92 ± 3.66 |