@A3shTnT 2024-12-11
> Check NVIDIA NSight Systems and look at the runtime for different kernel configurations. It may be that the variant you looked at with NSight Compute is faster but another one has become slower (though I wouldn't know why).

I will have a try.