@A3shTnT 2024-12-12
> > Check NVIDIA NSight Systems and look at the runtime for different kernel configurations. It may be that the variant you looked at with NSight Compute is faster but another one has become slower (though I wouldn't know why).
> 
> I'll give it a try.

Since this computer cannot run a model as large as Deepseek, I used the existing MAMBA model, but the problem should be quite similar. 
It seems that the uncontiguous concat operation has a relatively small proportion of execution time in the entire model, and the performance improvement is difficult to see. Perhaps updates targeting small operators may not achieve significant acceleration effects at the model level, and the impact of fluctuations in runtime function execution time may be even greater.I am not familiar with runtime, this is just my opinion.




**Kernel execution**

my implementation:
| Time (%) | Total Time (ns) | Instances | Avg (ns) | Med (ns) | Min (ns) | Max (ns) | StdDev (ns) | Name |
|----------|-----------------|-----------|----------|----------|----------|----------|-------------|------|
| 31.4     | 1508066774      | 16031     | 94071.9  | 59026.0  | 57390    | 955443   | 171105.2     | void mul_mat_vec |
|...|...|...|....|...|...|...|...|...|
| 1.1      | 51516703        | 31056     | 1658.8   | 963.0    | 898      | 74872    | 7035.6       | silu_f32                      |
|...|...|...|....|...|...|...|...|...|
| 0.4      | 20815533        | 144       | 144552.3 | 139977.0 | 136432   | 206558   | 16592.1      | void concat_f32_non_cont |
|...|...|...|....|...|...|...|...|...|


pr #10558(It can be considered as master, because changes are not related to concat):
| Time (%) | Total Time (ns) | Instances | Avg (ns) | Med (ns) | Min (ns) | Max (ns) | StdDev (ns) | Name |
|----------|-----------------|-----------|----------|----------|----------|----------|-------------|------|
| 31.2     | 1503470079      | 16031     | 93785.2  | 58678.0  | 57521    | 956623   | 170486.9     | void mul_mat_vec |
|...|...|...|....|...|...|...|...|...|
| 1.2      | 55916400        | 144       | 388308.3 | 387963.0 | 380733   | 399661   | 3754.4       | concat_f32_non_cont|
| 1.1      | 51843144        | 31056     | 1669.3   | 965.0    | 899      | 74970    | 7049.0       | silu_f32|
|...|...|...|....|...|...|...|...|...|


**Runtime function**
my implementation:
| Time (%) | Total Time (ns) | Num Calls | Avg (ns) | Med (ns) | Min (ns) | Max (ns) | StdDev (ns) | Name |
|----------|-----------------|-----------|----------|----------|----------|----------|-------------|------|
| 38.2     | 3830368168      | 6718      | 570165.0 | 17748.5  | 397      | 324782389| 8298624.3   | cudaStreamSynchronize |
| 25.2     | 2526634352      | 2027      | 1246489.6| 15925.0  | 1772     | 294406598| 8298396.3   | cuLibraryLoadData |
| 25.0     | 2507314668      | 360955    | 6946.3   | 5192.0   | 2242     | 3443261  | 15761.5     | cudaLaunchKernel |
| 6.2      | 617263233       | 3         | 205754411.0| 480931.0 | 180886   | 616601416| 355803975.0 | cudaMemGetInfo |
| 4.6      | 462075207       | 34389     | 13436.7  | 6838.0   | 3076     | 46874242 | 264276.4    | cudaMemcpyAsync |

pr #10558
| Time (%) | Total Time (ns) | Num Calls | Avg (ns) | Med (ns) | Min (ns) | Max (ns) | StdDev (ns) | Name |
|----------|-----------------|-----------|----------|----------|----------|----------|-------------|------|
| 43.6     | 3795485106      | 6718      | 564972.5 | 15840.5  | 342      | 313831670| 8389839.7   | cudaStreamSynchronize |
| 32.3     | 2811759210      | 360955    | 7789.8   | 4643.0   | 1991     | 4257846  | 18166.3     | cudaLaunchKernel |
| 14.8     | 1284635317      | 2027      | 633761.9 | 5042.0   | 1289     | 203751445| 5059546.5   | cuLibraryLoadData |
| 5.3      | 465147879       | 34389     | 13526.1  | 5898.0   | 2665     | 22423451 | 136640.3    | cudaMemcpyAsync |
| 3.1      | 271980756       | 3         | 90660252.0| 453325.0 | 181184   | 271346247| 156478720.9 | cudaMemGetInfo |