@A3shTnT 2024-12-12
> > Since this computer cannot run a model as large as Deepseek, I used the existing MAMBA model, but the problem should be quite similar.
> 
> You can profile the code from the command line with something like `nsys profile ./llama-bench`, then copy the resulting file to your desktop where you can analyze it. But these numbers already look very good so I would be fine with merging this as-is.

The numbers in the table are obtained through the nsight system, `nsys profile`, and then the summary table obtained through the `nsys stats` . The Nsight System is indeed a great tool for analyzing from the whole model overviews. Thank you for your advice.